Distribute aggregation tier logic to proper packages
- ParseISO8601Duration → utils/ (general-purpose time parser alongside
  NormalizeTimeToSpaceFormat and NormalizeDateTimeParam)
- ParseAggregationTiers, ResolveAggregationInterval, DefaultAggregationTiers,
  FallbackInterval → service/aggregation_tiers.go (domain logic used by
  ReadingsService)
- types/aggregation.go retains only pure types: AggregationTier struct,
  interval/function constants, response wrappers, error types

Tests follow their code: duration parser tests in utils/, tier tests
in service/.
